RELOAD: Widen the offset
This dissector has a lot of issues, and one of them is storing 32-bit
quantities retrieved from packet data into 16-bit integers. Widening
all the offsets fixes #21381, at least, though a lot more could be
done to clean this up.

(backported from commit 9f86e53c69a763ef34f2787301b482de1868738b)

+++ epan/dissectors/packet-reload.h
@@ -13,6 +13,6 @@
 #ifndef __PACKET_RELOAD_H__
 #define __PACKET_RELOAD_H__
 
-extern int dissect_reload_messagecontents(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree, uint16_t offset, uint16_t length);
+extern int dissect_reload_messagecontents(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree, unsigned offset, uint16_t length);
 
 #endif
